optimize non-prettyfied code paths for hash and array generation
The code to generate hashes and arrays has to check whether prettification attributes (like `indent`, `object_nl`, etc.) are being defined. Those are defined when setting up the `JSON_Generator_State` object, and never change afterwards during the lifetime of the object.

The compiler could optimize those checks, it it were able to understand that they don't change. However, this is hard or impossible to achieve, because the current code path can end up in the generic `generate_json` function, and, in `generate_json_object`, that even occurs via the `generate_json_i_object` callback.

In performance critical scenarios the JSON gem is not run with any prettification options. To improve the performance, this PR clones the existing `generate_json_object` and `generate_json_array` functions into `..._fast` versions, that skip all prettifications, and checks whether the pretty or the fast option should be applied when setting up the `JSON_Generator_State`.